README and bugs

+++ src/main/java/br/com/fellipeoliveira/meetingroom/gateways/impl/RoomSchedulingGatewayImpl.java
@@ -25,7 +25,7 @@ public class RoomSchedulingGatewayImpl implements RoomSchedulingGateway {
 
   @Override
   public void deleteScheduling(Long schedulingId) {
-    roomSchedulingRepository.deleteById(schedulingId.intValue());
+    roomSchedulingRepository.deleteById(schedulingId);
   }
 
   @Override
